src/main/java/seedu/address/logic/parser/AddAssignmentByTgCommandParser.java
package seedu.address.logic.parser;

import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
import static seedu.address.logic.parser.CliSyntax.PREFIX_ASSIGNMENT;
import static seedu.address.logic.parser.CliSyntax.PREFIX_DEADLINE;
import static seedu.address.logic.parser.CliSyntax.PREFIX_TUTORIAL_GROUP;

import java.util.stream.Stream;

import seedu.address.logic.commands.AddAssignmentByTgCommand;
import seedu.address.logic.parser.exceptions.ParseException;
import seedu.address.model.assignment.Assignment;
import seedu.address.model.assignment.AssignmentName;
import seedu.address.model.assignment.Deadline;
import seedu.address.model.assignment.Grade;
import seedu.address.model.assignment.Status;
import seedu.address.model.student.TutorialGroup;

/**
 * Parses input arguments and creates a new {@code AddAssignmentByTgCommand} object
 */
public class AddAssignmentByTgCommandParser implements Parser<AddAssignmentByTgCommand> {
    @Override
    public AddAssignmentByTgCommand parse(String args) throws ParseException {
        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args,
                PREFIX_TUTORIAL_GROUP, PREFIX_ASSIGNMENT, PREFIX_DEADLINE);
        argMultimap.verifyNoInvalidPrefixesFor(args);

        if (!arePrefixesPresent(argMultimap, PREFIX_TUTORIAL_GROUP, PREFIX_ASSIGNMENT, PREFIX_DEADLINE)
            || !argMultimap.getPreamble().isEmpty()) {
            throw new ParseException(
                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddAssignmentByTgCommand.MESSAGE_USAGE));
        }

        TutorialGroup tutorialGroup = ParserUtil.parseTutorialGroup(argMultimap.getValue(PREFIX_TUTORIAL_GROUP).get());
        AssignmentName assignmentName = ParserUtil.parseAssignmentName(argMultimap.getValue(PREFIX_ASSIGNMENT).get());
        Deadline deadline = ParserUtil.parseDeadline(argMultimap.getValue(PREFIX_DEADLINE).get());

        return new AddAssignmentByTgCommand(
                new Assignment(assignmentName, deadline, Status.getDefault(), Grade.getDefault()),
                tutorialGroup);
    }

    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
    }
}
